+++ src/idunno.AtProto/SecurityHelpers.cs
@@ -23,11 +23,18 @@ internal sealed class SecurityHelpers
     /// Implements simple SSRF validation on the specified <paramref name="uri"/> by checking if the host resolves to a public IP address.
     /// </summary>
     /// <param name="uri">The <see cref="Uri"/> to validate.</param>
+    /// <param name="allowInsecureProtocols">Indicates whether insecure protocols (e.g., HTTP) are allowed.</param>
+    /// <param name="allowLoopback">Indicates whether loopback addresses are allowed.</param>
     /// <param name="loggerFactory">The <see cref="ILoggerFactory"/> to use for logging any validation issues.</param>
     /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
     /// <returns><see langword="true" /> if the <paramref name="uri" /> is considered safe, otherwise <see langword="false"/>.</returns>
     /// <exception cref="ArgumentNullException">Thrown when <paramref name="uri"/> is <see langword="null"/>.</exception>
-    public static async Task<bool> DefaultDiscoveryUriValidator(Uri uri, ILoggerFactory? loggerFactory, CancellationToken cancellationToken = default)
+    public static async Task<bool> DefaultDiscoveryUriValidator(
+        Uri uri,
+        bool allowInsecureProtocols,
+        bool allowLoopback,
+        ILoggerFactory? loggerFactory,
+        CancellationToken cancellationToken = default)
     {
         ArgumentNullException.ThrowIfNull(uri);
 
@@ -35,7 +42,10 @@ internal sealed class SecurityHelpers
 
         ILogger<SecurityHelpers> logger = loggerFactory.CreateLogger<SecurityHelpers>();
 
-        if (Ssrf.IsUnsafeUri(uri))
+        if (Ssrf.IsUnsafeUri(
+            uri: uri,
+            allowInsecureProtocols: allowInsecureProtocols,
+            allowLoopback: allowLoopback))
         {
             Logger.UnsafeUri(logger, uri);
             return false;
@@ -45,7 +55,7 @@ internal sealed class SecurityHelpers
         {
             var ipAddress = IPAddress.Parse(uri.Host);
 
-            if (Ssrf.IsUnsafeIpAddress(ipAddress))
+            if (Ssrf.IsUnsafeIpAddress(ipAddress, allowLoopback))
             {
                 Logger.UnsafeIpAddress(logger, uri, ipAddress);
                 return false;
@@ -66,7 +76,7 @@ internal sealed class SecurityHelpers
 
             bool discoveredUnsafeIPAddress = false;
 
-            foreach (IPAddress ipAddress in hostEntry.AddressList.Where(Ssrf.IsUnsafeIpAddress))
+            foreach (IPAddress ipAddress in hostEntry.AddressList.Where(ip => Ssrf.IsUnsafeIpAddress(ip, allowLoopback)))
             {
                 Logger.UnsafeIpAddress(logger, uri, ipAddress);
                 discoveredUnsafeIPAddress = true;